Remove workarounds for TypeScript in Folder component HTML
In an earlier implementation, a Treetop node's 'children' and 'url'
properties were optional. Because the non-null assertion operator (!)
cannot be used in Svelte templates [1], accessing these properties was
refactored out of the template.

This workaround is no longer necessary now that the Treetop node types
are distinguished through a TypeScript discriminated union. Therefore,
remove them.

[1] https://github.com/sveltejs/svelte/issues/4701