Prefer contextMenus API over menus API
Replace usage of the menus API with the contextMenus API. This is a step
towards compatibility with Chrome.

Note that the contextMenus API lacks some features of the menus API.
Specifically, the click data doesn't include a targetElementId property,
like in menus.OnClickData [1]. Therefore, we now manually track the
target element of the context menu.

[1] https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/API/menus/OnClickData