Fix bookmark last visit time on Chrome
The `chrome.history.getVisits` API behaves differently on different
browsers: Firefox sorts the visits in reverse chronological order and
Chrome sorts the visits in chronological order.

Update the HistoryManager to handle both cases. This fixes a bug on
Chrome where a previously visited bookmark might not display in a larger
font even if it's been visited recently.